requirements-dev.txt
pytest
pylama
mypy
types-requests
isort
black
flake8
uvicorn
lxml-stubs
